Move ploting into wandb.py
